Add space at the end of page
In order to test scrolls, or space in the page, we may want to position
the selectic component at the top of the page.
An easy way is to add space at the end of the page and then to scroll
the page to put the Selectic component at top.